LG-2715: re-arrange classes
